requirements.txt
# ============================================
# 生产环境依赖
# ============================================

# DeepSeek AI API 客户端
# 用于调用 DeepSeek AI 进行题目分析
openai>=1.0.0

# 自动化控制库
# 用于模拟鼠标点击和键盘操作
pyautogui>=0.9.54

# 图像处理库
# 用于截图和图像格式转换
pillow>=10.0.0

# HTTP 请求库
# 用于调用百度 OCR API
requests>=2.31.0

# ============================================
# 开发和测试依赖
# ============================================

# 测试框架
# 用于运行单元测试和集成测试
pytest>=7.0.0

# 测试覆盖率工具
# 用于生成代码覆盖率报告
pytest-cov>=4.0.0

# 测试 Mock 工具
# 用于模拟外部依赖（API 调用等）
pytest-mock>=3.0.0

# ============================================
# 可选依赖（用于打包）
# ============================================

# PyInstaller - 用于打包成可执行文件
# 如需打包，请取消下面的注释
# pyinstaller>=5.0.0

